fix(Mutiple Chat Rooms): resolve multiple chat rooms bug when user request (#110)
* fix(Mutiple Chat Rooms): resolve multiple chat rooms bug when user request

This patch resolves the bug where the user placed multiple offers that results
in multiple chat rooms and conversation queries whenever these data are requested.
The issue is resolved by outerjoining the match table. Test cases are also setup
to prevent this issue in future.

* fix(Test Cases): Broke test case

This patch fixes the test case broke in test_match_services. This
is due to the addtional field match_id.

Note that the buyer_id and seller_id in chat_rooms are not removed
because the author does not feel confident.

* fix(Chat Tests): Resolve comments

* resolve comments