Fix db error (#111)
Reason for changing the setup_db.sh
Run the script will not work because the migration scripts for friendly name relies on the db model. Running this script causes me confusion because I did not expect the migrations to be setup that way

This patch separates the commands into three separate files. setup_db.sh drop and create the database, as well as creating the role for the db. The run_migrations.sh is responsible for upgrading the database. run_seeds.sh is responsible for seeding the database.

There is also one more file added run_downgrade.sh. It is responsible for downgrading the database to 1 migration behind. I needed this badly for this patch.

Note that the foreign key name is included so that you can downgrade and upgrade easily.